Fixed barchart to work with new Chart JS + Vue3
