+++ src/components/campaigns/campaign_building_list.vue
@@ -13,16 +13,16 @@
     <el-row class="buildingScroll" v-loading="!loaded" element-loading-background="rgba(0, 0, 0, 0.8)">
       <el-row class="buildingRow" v-if="!loaded"> &nbsp; </el-row>
       <el-row class="buildingRow" v-for="block in blocks" :key="block.path" ref="buildingRows">
-        <el-col v-if="loaded" :class="[value === block.path ? 'buildingCol selected' : 'buildingCol']" :span="24">
-          <div :class="[value === block.path ? 'outerClip selected' : 'outerClip']">
+        <el-col v-if="loaded" :class="[modelValue === block.path ? 'buildingCol selected' : 'buildingCol']" :span="24">
+          <div :class="[modelValue === block.path ? 'outerClip selected' : 'outerClip']">
             <div
               v-if="!isNaN(accumulatedPercentage(block.path))"
-              :class="[value === block.path ? 'innerClip selected' : 'innerClip']"
+              :class="[modelValue === block.path ? 'innerClip selected' : 'innerClip']"
               :style="`background-color:${computedColor(block.path)};`"
               @click="buildingClick(block.path)"
             >
               <i class="fas fa-trophy" v-if="place(block.path) <= 3 && place(block.path) >= 1"
-                ><span :class="[value === block.path ? 'innerTrophy selected' : 'innerTrophy']">{{
+                ><span :class="[modelValue === block.path ? 'innerTrophy selected' : 'innerTrophy']">{{
                   place(block.path)
                 }}</span></i
               >
@@ -37,11 +37,11 @@
             <!-- Display "No Data" on block, remove button functionality of block if NaN percentage detected-->
             <div
               v-else
-              :class="[value === block.path ? 'innerClip selected' : 'innerClipNoData']"
+              :class="[modelValue === block.path ? 'innerClip selected' : 'innerClipNoData']"
               :style="`background-color:${computedColor(block.path)};`"
             >
               <i class="fas fa-trophy" v-if="place(block.path) <= 3 && place(block.path) >= 1"
-                ><span :class="[value === block.path ? 'innerTrophy selected' : 'innerTrophy']">{{
+                ><span :class="[modelValue === block.path ? 'innerTrophy selected' : 'innerTrophy']">{{
                   place(block.path)
                 }}</span></i
               >
@@ -55,7 +55,7 @@
 </template>
 <script>
 export default {
-  props: ['path', 'loaded', 'value'],
+  props: ['path', 'loaded', 'modelValue'],
   data () {
     return {
       activePath: null
@@ -109,10 +109,10 @@ export default {
   },
   methods: {
     buildingClick: function (building) {
-      if (this.value === building) {
-        this.$emit('input', this.path + '/block_default')
+      if (this.modelValue === building) {
+        this.$emit('update:modelValue', this.path + '/block_default')
       } else {
-        this.$emit('input', building)
+        this.$emit('update:modelValue', building)
       }
     },
     accumulatedPercentage: function (path) {
